requirements.txt
# 核心依赖
PyExecJS>=1.5.1
requests>=2.31.0
loguru>=0.7.0
python-dotenv>=1.0.0
retry>=0.9.2
openpyxl>=3.1.0

# 数据验证
pydantic>=2.0.0

# CLI工具
click>=8.1.0

# 性能监控
psutil>=5.9.0